fix(suite-desktop): refactor desktop updater and prioritize updater modal
